feat: Fault Testing Support - update client to not rely on Optimism dependencies
