+++ client/client.go
@@ -7,7 +7,7 @@ import (
 	"io"
 	"net/http"
 
-	"github.com/Layr-Labs/eigenda-proxy/server"
+	"github.com/Layr-Labs/eigenda-proxy/common"
 )
 
 // TODO: Add support for custom http client option
@@ -19,7 +19,7 @@ type Config struct {
 // ProxyClient is an interface for communicating with the EigenDA proxy server
 type ProxyClient interface {
 	Health() error
-	GetData(ctx context.Context, cert []byte, domain server.DomainType) ([]byte, error)
+	GetData(ctx context.Context, cert []byte, domain common.DomainType) ([]byte, error)
 	SetData(ctx context.Context, b []byte) ([]byte, error)
 }
 
@@ -60,7 +60,7 @@ func (c *client) Health() error {
 }
 
 // GetData fetches blob data associated with a DA certificate
-func (c *client) GetData(ctx context.Context, comm []byte, domain server.DomainType) ([]byte, error) {
+func (c *client) GetData(ctx context.Context, comm []byte, domain common.DomainType) ([]byte, error) {
 	url := fmt.Sprintf("%s/get/0x%x?domain=%s&commitment_mode=simple", c.cfg.URL, comm, domain.String())
 
 	if c.cfg.Actor != "" {
