+++ server/domain_type.go
@@ -1,55 +1,20 @@
 package server
 
 import (
-	"fmt"
 	"net/http"
+
+	"github.com/Layr-Labs/eigenda-proxy/common"
 )
 
-var (
-	ErrInvalidDomainType = fmt.Errorf("invalid domain type")
-)
-
-// DomainType is a enumeration type for the different data domains for which a
-// blob can exist between
-type DomainType uint8
-
-const (
-	BinaryDomain DomainType = iota
-	PolyDomain
-	UnknownDomain
-)
-
-func (d DomainType) String() string {
-	switch d {
-	case BinaryDomain:
-		return "binary"
-	case PolyDomain:
-		return "polynomial"
-	default:
-		return "unknown"
-	}
-}
-
-func StrToDomainType(s string) DomainType {
-	switch s {
-	case "binary":
-		return BinaryDomain
-	case "polynomial":
-		return PolyDomain
-	default:
-		return UnknownDomain
-	}
-}
-
-func ReadDomainFilter(r *http.Request) (DomainType, error) {
+func ReadDomainFilter(r *http.Request) (common.DomainType, error) {
 	query := r.URL.Query()
 	key := query.Get(DomainFilterKey)
 	if key == "" { // default
-		return BinaryDomain, nil
+		return common.BinaryDomain, nil
 	}
-	dt := StrToDomainType(key)
-	if dt == UnknownDomain {
-		return UnknownDomain, ErrInvalidDomainType
+	dt := common.StrToDomainType(key)
+	if dt == common.UnknownDomain {
+		return common.UnknownDomain, common.ErrInvalidDomainType
 	}
 
 	return dt, nil
